requirements.txt
# Backend API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-socketio>=5.10.0
websockets>=12.0

# AI & ML
openai>=1.50.0
numpy>=1.24.0

# Knowledge Graph
networkx>=3.0

# Utilities
python-dotenv>=1.0.0

# Audio (optional)
sounddevice>=0.4.6
soundfile>=0.12.1
rich>=13.7.0


